tests/wpt/tests/html/semantics/invokers/invoketarget-on-popover-invalid-behavior.tentative.html

<!doctype html>
<meta charset="utf-8" />
<meta name="author" title="Keith Cirkel" href="mailto:wpt@keithcirkel.co.uk" />
<meta name="timeout" content="long" />
<link rel="help" href="https://open-ui.org/components/invokers.explainer/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/invoker-utils.js"></script>

<div id="invokee" popover>
  <button id="containedinvoker" invoketarget="invokee"></button>
</div>
<button id="invokerbutton" invoketarget="invokee"></button>

<script>
  function resetState() {
    invokerbutton.removeAttribute("invokeaction");
    containedinvoker.removeAttribute("invokeaction");
    try {
      invokee.hidePopover();
    } catch {}
    invokee.setAttribute("popover", "");
  }

  // invalid actions on showpopover
  ["foo-bar", "showmodal", "showpicker", "open", "close"].forEach((action) => {
    promise_test(async function (t) {
      t.add_cleanup(resetState);
      invokerbutton.invokeAction = action;
      assert_false(invokee.matches(":popover-open"));
      await clickOn(invokerbutton);
      assert_false(invokee.matches(":popover-open"));
    }, `invoking (as ${action}) on popover does nothing`);

    promise_test(async function (t) {
      t.add_cleanup(resetState);
      invokerbutton.invokeAction = action;
      invokee.showPopover();
      assert_true(invokee.matches(":popover-open"));
      await clickOn(invokerbutton);
      assert_true(invokee.matches(":popover-open"));
    }, `invoking (as ${action}) on open popover does nothing`);
  });
</script>
